refactor(nsa.c): euler's method return type
use double instead of float for more precision in approxResultByEuler()

+++ core/source/nsa.h
@@ -2,5 +2,5 @@
 
 double getRootByBisection(double (*f)(double value), int accuracyAfterDecimalPoint);
 double getRootByRegulaFalsi(double (*f)(double value), int accuracyAfterDecimalPoint);
-float approxResultByEuler(float x, float y, float lastValueOfX, float stepSize, float (*f)(float x, float y), int accuracyAfterDecimalPoint);
+double approxResultByEuler(double x, double y, double lastValueOfX, double stepSize, double (*f)(double x, double y), unsigned short int accuracyAfterDecimalPoint);
 double getRootBySecant(double (*f)(double x), unsigned short int accuracyFactor);
